requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
dj-database-url
Django
django-autoslug
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-jazzmin
django-tinymce
djangorestframework
idna
Markdown
pillow
psycopg2
psycopg2-binary
python-decouple
pytz
requests
six
sqlparse
sslcommerz
sslcommerz-lib
typing_extensions
urllib3
